refactor:  More circular reference resoltution.
Puts AsyncIterator in models.utils, which can be safely imported by both snek and discord models.
