+++ dis_snek/models/__init__.py
@@ -1,2 +1,3 @@
 from .discord import *
 from .snek import *
+from .utils import *
