+++ dis_snek/models/discord/reaction.py
@@ -6,7 +6,7 @@ from dis_snek.client.const import MISSING
 from dis_snek.client.utils.attr_utils import define, field
 from dis_snek.models.discord.emoji import PartialEmoji
 from dis_snek.models.discord.snowflake import to_snowflake
-from dis_snek.models.snek.iterator import AsyncIterator
+from dis_snek.models.utils.iterator import AsyncIterator
 from .base import ClientObject
 
 if TYPE_CHECKING:
