+++ dis_snek/models/utils/__init__.py
@@ -1,0 +1,3 @@
+from .iterator import AsyncIterator
+
+__all__ = ["AsyncIterator"]
